requirements.txt
numpy>=1.24
pandas>=2.1
scipy>=1.11
scikit-learn>=1.3
statsmodels>=0.14
pyyaml>=6.0
pyarrow>=14
rich>=13
# optional dev tools
flake8>=6
black>=23
pytest>=7
